# perf: Add fast paths for series.arg_sort and dataframe.sort