--- a/crates/polars-core/src/chunked_array/ops/sort/arg_sort.rs
+++ b/crates/polars-core/src/chunked_array/ops/sort/arg_sort.rs
@@ -15,18 +15,34 @@ where
         options.multithreaded,
     );
 }
-fn reverse_stable_no_nulls<I, J, T>(iters: I, len: usize) -> Vec<u32>
+// Compute the indexes after reversing a sorted array, maintaining
+// the order of equal elements, in linear time. Faster than sort_impl
+//  as we avoid allocating extra memory.
+pub(super) fn reverse_stable_no_nulls<I, J, T>(iters: I, len: usize) -> Vec<IdxSize>
 where
     I: IntoIterator<Item = J>,
     J: IntoIterator<Item = T>,
     T: TotalOrd + Send + Sync,
 {
-    let mut current_start = 0;
-    let mut current_end = 1;
+    let mut current_start: IdxSize = 0;
+    let mut current_end: IdxSize = 1;
     let mut flattened_iter = iters.into_iter().flatten();
     let first_element = flattened_iter.next();
-    let mut rev_idx = Vec::with_capacity(len);
+    let mut rev_idx: Vec<IdxSize> = Vec::with_capacity(len);
     let mut i: IdxSize;
+    // We traverse the array , comparing consecutive elements.
+    // We maintain the start and end indice of elements with same value.
+    // When we see a new element we push the previous indices in reverse order.
+    // We do a final reverse to get stable reverse index.
+    // Example -
+    // 1 2 2 3 3 3 4
+    // 0 1 2 3 4 5 6
+    // We get start and end position of equal values -
+    // 0 1-2 3-5 6
+    // We insert the indexes of equal elements in reverse
+    // 0 2 1 5 4 3 6
+    // Then do a final reverse
+    // 6 3 4 5 1 2 0
     match first_element {
         Some(value) => {
             let mut previous_element = value;
@@ -34,10 +50,11 @@ where
                 if current_element.tot_cmp(&previous_element) == Ordering::Equal {
                     current_end += 1;
                 } else {
-                    //rev_idx.extend((current_start..current_end).rev());
+                    // Insert in reverse order
                     i = current_end;
                     while i > current_start {
                         i = i - 1;
+                        //SAFETY - we allocated enough
                         unsafe { rev_idx.push_unchecked(i) };
                     }
                     current_start = current_end;
@@ -45,7 +62,12 @@ where
                 }
                 previous_element = current_element;
             }
-            rev_idx.extend((current_start..current_end).rev());
+            i = current_end;
+            while i > current_start {
+                i = i - 1;
+                unsafe { rev_idx.push_unchecked(i) };
+            }
+            // Final reverse
             rev_idx.reverse();
             return rev_idx;
         },
@@ -71,6 +93,9 @@ where
     let nulls_last = options.nulls_last;
     let null_cap = if nulls_last { null_count } else { len };
 
+    // Fast path
+    // Only if array is already sorted in the required ordered and
+    // nulls are also in the correct position
     if (options.descending && is_sorted_descending_flag)
         || (!options.descending && is_sorted_ascending_flag)
     {
@@ -169,6 +194,9 @@ where
     J: IntoIterator<Item = T>,
     T: TotalOrd + Send + Sync,
 {
+    // Fast path
+    // 1) If array is already sorted in the required ordered .
+    // 2) If array is reverse sorted -> we do a stable reverse.
     if (options.descending && is_sorted_descending_flag)
         || (!options.descending && is_sorted_ascending_flag)
     {
@@ -244,4 +272,60 @@ pub(crate) fn arg_sort_row_fmt(
 
     let ca: NoNull<IdxCa> = items.into_iter().map(|tpl| tpl.0).collect();
     Ok(ca.into_inner())
+}
+#[cfg(test)]
+mod test {
+    use sort::arg_sort::reverse_stable_no_nulls;
+
+    use crate::prelude::*;
+
+    #[test]
+    fn test_reverse_stable_no_nulls() {
+        let a = Int32Chunked::new(
+            PlSmallStr::from_static("a"),
+            &[
+                Some(1), // 0
+                Some(2), // 1
+                Some(2), // 2
+                Some(3), // 3
+                Some(3), // 4
+                Some(3), // 5
+                Some(4), // 6
+            ],
+        );
+        let idx = reverse_stable_no_nulls(a.into_iter(), 7);
+        let expected = [6, 3, 4, 5, 1, 2, 0];
+        assert_eq!(idx, expected);
+
+        let a = Int32Chunked::new(
+            PlSmallStr::from_static("a"),
+            &[
+                Some(1), // 0
+                Some(2), // 1
+                Some(3), // 2
+                Some(4), // 3
+                Some(5), // 4
+                Some(6), // 5
+                Some(7), // 6
+            ],
+        );
+        let idx = reverse_stable_no_nulls(a.into_iter(), 7);
+        let expected = [6, 5, 4, 3, 2, 1, 0];
+        assert_eq!(idx, expected);
+
+        let a = Int32Chunked::new(
+            PlSmallStr::from_static("a"),
+            &[
+                Some(1), // 0
+            ],
+        );
+        let idx = reverse_stable_no_nulls(a.into_iter(), 1);
+        let expected = [0];
+        assert_eq!(idx, expected);
+
+        let empty_array: [i32; 0] = [];
+        let a = Int32Chunked::new(PlSmallStr::from_static("a"), &empty_array);
+        let idx = reverse_stable_no_nulls(a.into_iter(), 0);
+        assert_eq!(idx.len(), 0);
+    }
 }