--- a/crates/polars-core/src/chunked_array/ops/sort/arg_sort.rs
+++ b/crates/polars-core/src/chunked_array/ops/sort/arg_sort.rs
@@ -290,7 +290,7 @@ mod test {
                 Some(4), // 6
             ],
         );
-        let idx = reverse_stable_no_nulls(a.into_iter(), 7);
+        let idx = reverse_stable_no_nulls(&a, 7);
         let expected = [6, 3, 4, 5, 1, 2, 0];
         assert_eq!(idx, expected);
 
@@ -306,7 +306,7 @@ mod test {
                 Some(7), // 6
             ],
         );
-        let idx = reverse_stable_no_nulls(a.into_iter(), 7);
+        let idx = reverse_stable_no_nulls(&a, 7);
         let expected = [6, 5, 4, 3, 2, 1, 0];
         assert_eq!(idx, expected);
 
@@ -316,13 +316,13 @@ mod test {
                 Some(1), // 0
             ],
         );
-        let idx = reverse_stable_no_nulls(a.into_iter(), 1);
+        let idx = reverse_stable_no_nulls(&a, 1);
         let expected = [0];
         assert_eq!(idx, expected);
 
         let empty_array: [i32; 0] = [];
         let a = Int32Chunked::new(PlSmallStr::from_static("a"), &empty_array);
-        let idx = reverse_stable_no_nulls(a.into_iter(), 0);
+        let idx = reverse_stable_no_nulls(&a, 0);
         assert_eq!(idx.len(), 0);
     }
 }