--- a/crates/polars-core/src/chunked_array/ops/sort/categorical.rs
+++ b/crates/polars-core/src/chunked_array/ops/sort/categorical.rs
@@ -67,6 +67,9 @@ impl CategoricalChunked {
                 options,
                 self.physical().null_count(),
                 self.len(),
+                false,
+                false,
+                false,
             )
         } else {
             self.physical().arg_sort(options)