--- a/crates/polars-core/src/chunked_array/ops/sort/categorical.rs
+++ b/crates/polars-core/src/chunked_array/ops/sort/categorical.rs
@@ -67,8 +67,7 @@ impl CategoricalChunked {
                 options,
                 self.physical().null_count(),
                 self.len(),
-                false,
-                false,
+                IsSorted::Not,
                 false,
             )
         } else {