--- a/crates/polars-core/src/chunked_array/ops/sort/mod.rs
+++ b/crates/polars-core/src/chunked_array/ops/sort/mod.rs
@@ -137,9 +137,9 @@ macro_rules! sort_with_fast_path {
             // there are nulls
             if $ca.null_count() > 0 {
                 // if the nulls are already last we can clone
-                if ($options.nulls_last && $ca.get($ca.len() - 1).is_none())  ||
+                if $options.nulls_last && $ca.get($ca.len() - 1).is_none()  ||
                 // if the nulls are already first we can clone
-                $ca.get(0).is_none()
+                (!$options.nulls_last && $ca.get(0).is_none())
                 {
                     return $ca.clone();
                 }