Add fast paths for series.arg_sort and dataframe.sort(single column) in the following cases -
1) Key column has no nulls and is sorted in the required order 
2) Key column has no nulls and is sorted in the opposite order 
3) Key column has nulls ,is sorted in the required order and has nulls in the correct place

For the second case, if maintain_order is false, technically we can simply return the reverse of 0..len. However current behavior is that we still do a stable sort on it, maintaining order of elements. This is also the expectation in some unrelated test cases. To maintain this behavior, I have implemented a linear algorithm to do a stable reverse sort. This is faster than the current implementation which is also linear for sorting already sorted data, but takes more time due to creating a copy of the data.

Closes #19364
### Benchmarks
**dataframe.sort**
<details>
<summary>Code </summary>

```
import numpy as np
import polars as pl
n = 1_000_000
df = pl.DataFrame({
    "a": np.random.rand(n),
    "b": np.random.rand(n),
})

# -- One column
print("A unsorted")
df2 = df.select("a")
%timeit df2.sort("a")

print("A sorted")
df2 = df.select("a").sort("a")
%timeit df2.sort("a")

# -- Two columns
print("A unsorted, B unsorted")
df2 = df.clone()
%timeit df2.sort("a")

print("A sorted, B unsorted")
df2 = df.sort("a")
%timeit df2.sort("a")

print("A sorted reverse, B unsorted")
df2 = df.sort("a",descending=True)
%timeit df2.sort("a")


df_nulls = pl.DataFrame({
    "a": np.random.rand(n).tolist()+[None],
    "b": np.random.rand(n).tolist()+[None],
})
print("A sorted, with null")
df2_nulls = df_nulls.sort("a")
%timeit df2_nulls.sort("a")
```
</details>

| case  | Time taken Before| Time taken After |
| ------------- | ------------- | ------------- |
| A unsorted| 32.7 ms ± 651 μs | 32 ms ± 1.33 ms |
| A sorted| 11.6 ms ± 52.6 μs | **6.08 μs ± 24.2 ns** |
| A sorted reverse| 17.6 ms ± 1.43 ms | **4.75 ms ± 244 μs** |
| A sorted, with null| 14.1 ms ± 560 μs| **6.14 μs ± 43.5 ns** |

**series.arg_sort**
<details>
<summary> Code </summary>

```
import numpy as np
#import polars as pl
n = 1_000_000
df = pl.DataFrame({
    "a": np.random.rand(n),
    "b": np.random.rand(n),
})
s=df["a"]
%timeit s.arg_sort()

s=df.sort("a")["a"]
%timeit s.arg_sort()
%timeit s.arg_sort(descending=True)
df_nulls = pl.DataFrame({
    "a": np.random.rand(n).tolist()+[None],
    "b": np.random.rand(n).tolist()+[None],
})
s_nulls=df_nulls.sort("a")["a"]
%timeit s_nulls.arg_sort()
```
</details>

| case  | Time taken Before| Time taken After |
| ------------- | ------------- | ------------- |
| A unsorted| 27.2 ms ± 715 μs| 27.5 ms ± 472 μs |
| A sorted| 9.85 ms ± 51.2 μs | **71.2 μs ± 362 ns** |
| A sorted reverse| 12.7 ms ± 298 μs | **1.99 ms ± 9.96 μs** |
| A sorted, with null| 11 ms ± 67.7 μs | **69.8 μs ± 744 ns** |

### Testing

We need to ensure that even if we take fast path we do not change the final array. I have modified some pre existing tests to also test for fast path. I have added an additional test which checks for correctness of sorting.